Linenumber directives from markdown files
With this patch you get line directives (common and golang-specific) in
your output. Useful when debugging, by rightclicking on acme or with gF
in vim. Typical output with `go run' or `go build':

  # command-line-arguments
  WhitespacePreservation.md:38: invalid character U+0023 '#'
  WhitespacePreservation.md:38: undefined: TEST
  WhitespacePreservation.md:39: missing return at end of function
  LineNumbers.md:147: undefined: parseHeader

This is could be seen as an improvment, instead of finding a bug in your
go code and then trying to find the right place in your md-files. It is
however very verbose (in your output), it is however "easy" to remove
the directives when you want to publish your go-code:

  sed -i '/^\/\/line/d' main.go