Expand the markup
By adding ~ to ` for allowed characters in code fences and ending code
blocks with as many or more characters as we start a codeblock with, we
are probably more conformant with other markdown processors and allows
us to create lmt acceptable markdown which produces markdown.

Moreover macros don't just start with <<< but also //<. This helps by
making it possible to run gofmt on codeblocks in your favourite editor.