Bug when using lmt on code in "unknown" languages
When removing a default case for languages without a small guard for
empty strings in fmt.Println, a tiny bug was introduced which pasted
lots of junk in our inputfiles. Some of it was not parseable perl.

This commit tries to correct it, cleans up the code a wee bit and tries
to make it easier to add more languages in the future.