Support C++ language blocks
\w is a synonym for "[0-9A-Za-z_]" in a regular expression.

C++ does not match that regex, because of the "+". Add "+"
to the language part of the regexes explicitly, so that C++
language code blocks can be extracted.

According to the languages list in
https://www.rubycoloredglasses.com/2013/04/languages-supported-by-github-flavored-markdown/
"+" should be the only special non-word character needed to
support the languages supported by GitHub flavored markdown.